카테고리별로 불러오기

+++ src/main/webapp/javascript/main/event.js
@@ -1,14 +1,18 @@
 function addButtonClickEvent() {
 	var moreButton = document.querySelector(".more > .btn");
-	moreButton.addEventListener('click', function(event){
+	moreButton.addEventListener("click", function(event){
 		sendGetThumbnailsAjax();
 	});
 }
 
 function addCategoryClickEvent() {
-	var Category = document.querySelector(".section_event_tab > .event_tab_lst");
-	moreButton.addEventListener('click', function(event){
+	var categoryButton = document.querySelector(".section_event_tab > .event_tab_lst");
+	categoryButton.addEventListener("click", function(event){
+		var categoryId = event.target.id.replace("category_", "");
+		if (event.target.tagName != "SPAN") {
+			return;
+		}
 		deleteAllThumbnail();
-		sendGetThumbnailsAjax();
+		sendGetThumbnailsAjax(categoryId);
 	});
 }
